+++ src/digirock/typing.py
@@ -1,6 +1,6 @@
 from typing import Union
-from numpy.typing import NDArray
+from numpy import ndarray
 import pathlib
 
-NDArrayOrFloat = Union[NDArray, float]
+NDArrayOrFloat = Union[ndarray, float]
 Pathlike = Union[str, pathlib.Path]
